backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
werkzeug==3.0.1
pandas==2.2.0
openpyxl==3.1.2
pyxlsb==1.0.10
